# mobile/openapi.json
{
 "openapi": "3.1.0",
 "info": {
  "title": "Triplet API",
  "description": "Trip planning aplication API",
  "version": "1.0.0"
 },
 "paths": {
...
  "/trips/{trip_id}": {
   "get": {
    "tags": [
     "Trips"
    ],
    "summary": "Get Trip",
    "operationId": "get_trip_trips__trip_id__get",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "parameters": [
     {
      "name": "trip_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Trip Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/TripResponse"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   },
   "patch": {
    "tags": [
     "Trips"
    ],
    "summary": "Update Trips",
    "operationId": "update_trips_trips__trip_id__patch",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "parameters": [
     {
      "name": "trip_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Trip Id"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/TripUpdate"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/TripResponse"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   },
   "delete": {
    "tags": [
     "Trips"
    ],
    "summary": "Delete Trip",
    "operationId": "delete_trip_trips__trip_id__delete",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "parameters": [
     {
      "name": "trip_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Trip Id"
      }
     }
    ],
    "responses": {
     "204": {
      "description": "Successful Response"
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
  "/trips/{trip_id}/members": {
   "get": {
    "tags": [
     "Trip Members"
    ],
    "summary": "Get Members",
    "operationId": "get_members_trips__trip_id__members_get",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "parameters": [
     {
      "name": "trip_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Trip Id"
      }
     },
     {
      "name": "limit",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "maximum": 100,
       "minimum": 1,
       "default": 50,
       "title": "Limit"
      }
     },
     {
      "name": "offset",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "minimum": 0,
       "default": 0,
       "title": "Offset"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "array",
         "items": {
          "$ref": "#/components/schemas/MemberResponse"
         },
         "title": "Response Get Members Trips  Trip Id  Members Get"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/trips/{trip_id}/members/{user_id}": {
   "patch": {
    "tags": [
     "Trip Members"
    ],
    "summary": "Update Member Role",
    "operationId": "update_member_role_trips__trip_id__members__user_id__patch",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "parameters": [
     {
      "name": "trip_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Trip Id"
      }
     },
     {
      "name": "user_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "User Id"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/MemberRoleUpdate"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/MemberResponse"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   },
   "delete": {
    "tags": [
     "Trip Members"
    ],
    "summary": "Remove Member",
    "operationId": "remove_member_trips__trip_id__members__user_id__delete",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "parameters": [
     {
      "name": "trip_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Trip Id"
      }
     },
     {
      "name": "user_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "User Id"
      }
     }
    ],
    "responses": {
     "204": {
      "description": "Successful Response"
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/trips/{trip_id}/invitations": {
   "post": {
    "tags": [
     "Trip Members"
    ],
    "summary": "Create Invitation",
    "operationId": "create_invitation_trips__trip_id__invitations_post",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "parameters": [
     {
      "name": "trip_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Trip Id"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/InvitationCreate"
       }
      }
     }
    },
    "responses": {
     "201": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/InvitationResponse"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   },
   "get": {
    "tags": [
     "Trip Members"
    ],
    "summary": "Get Trip Invitations",
    "operationId": "get_trip_invitations_trips__trip_id__invitations_get",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "parameters": [
     {
      "name": "trip_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Trip Id"
      }
     },
     {
      "name": "limit",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "maximum": 100,
       "minimum": 1,
       "default": 50,
       "title": "Limit"
      }
     },
     {
      "name": "offset",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "minimum": 0,
       "default": 0,
       "title": "Offset"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "array",
         "items": {
          "$ref": "#/components/schemas/InvitationResponse"
         },
         "title": "Response Get Trip Invitations Trips  Trip Id  Invitations Get"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/trips/{trip_id}/invitations/{invitation_id}": {
   "delete": {
    "tags": [
     "Trip Members"
    ],
    "summary": "Cancel Invitation",
    "operationId": "cancel_invitation_trips__trip_id__invitations__invitation_id__delete",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "parameters": [
     {
      "name": "trip_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Trip Id"
      }
     },
     {
      "name": "invitation_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Invitation Id"
      }
     }
    ],
    "responses": {
     "204": {
      "description": "Successful Response"
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/invitations": {
   "get": {
    "tags": [
     "Invitations"
    ],
    "summary": "Get My Invitations",
    "operationId": "get_my_invitations_invitations_get",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "parameters": [
     {
      "name": "limit",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "maximum": 100,
       "minimum": 1,
       "default": 50,
       "title": "Limit"
      }
     },
     {
      "name": "offset",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "minimum": 0,
       "default": 0,
       "title": "Offset"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "array",
         "items": {
          "$ref": "#/components/schemas/InvitationResponse"
         },
         "title": "Response Get My Invitations Invitations Get"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
 },
 "components": {
  "schemas": {
...
   "Destination": {
    "properties": {
     "name": {
      "type": "string",
      "maxLength": 120,
      "minLength": 1,
      "title": "Name"
     },
     "address": {
      "anyOf": [
       {
        "type": "string",
        "maxLength": 255
       },
       {
        "type": "null"
       }
      ],
      "title": "Address"
     },
     "latitude": {
      "anyOf": [
       {
        "type": "number",
        "maximum": 90.0,
        "minimum": -90.0
       },
       {
        "type": "null"
       }
      ],
      "title": "Latitude"
     },
     "longitude": {
      "anyOf": [
       {
        "type": "number",
        "maximum": 180.0,
        "minimum": -180.0
       },
       {
        "type": "null"
       }
      ],
      "title": "Longitude"
     },
     "country_code": {
      "anyOf": [
       {
        "type": "string",
        "maxLength": 2,
        "minLength": 2
       },
       {
        "type": "null"
       }
      ],
      "title": "Country Code"
     }
    },
    "type": "object",
    "required": [
     "name"
    ],
    "title": "Destination",
    "description": "One place a trip goes. A pin and country are filled in by the server when it can."
   },
...
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "type": "array",
      "title": "Detail"
     }
    },
    "type": "object",
    "title": "HTTPValidationError"
   },
   "InvitationCreate": {
    "properties": {
     "email": {
      "anyOf": [
       {
        "type": "string",
        "format": "email"
       },
       {
        "type": "null"
       }
      ],
      "title": "Email"
     },
     "user_id": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "User Id"
     }
    },
    "type": "object",
    "title": "InvitationCreate"
   },
   "InvitationResponse": {
    "properties": {
     "id": {
      "type": "integer",
      "title": "Id"
     },
     "trip_id": {
      "type": "integer",
      "title": "Trip Id"
     },
     "user_id": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "User Id"
     },
     "invited_by_id": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Invited By Id"
     },
     "status": {
      "type": "string",
      "title": "Status"
     },
     "created_at": {
      "type": "string",
      "format": "date-time",
      "title": "Created At"
     },
     "trip_title": {
      "type": "string",
      "title": "Trip Title"
     },
     "trip_destination": {
      "type": "string",
      "title": "Trip Destination"
     },
     "trip_start_date": {
      "anyOf": [
       {
        "type": "string",
        "format": "date"
       },
       {
        "type": "null"
       }
      ],
      "title": "Trip Start Date"
     },
     "trip_end_date": {
      "anyOf": [
       {
        "type": "string",
        "format": "date"
       },
       {
        "type": "null"
       }
      ],
      "title": "Trip End Date"
     },
     "invitee_name": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Invitee Name"
     },
     "invitee_email": {
      "type": "string",
      "format": "email",
      "title": "Invitee Email"
     },
     "invited_by_name": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Invited By Name"
     }
    },
    "type": "object",
    "required": [
     "id",
     "trip_id",
     "status",
     "created_at",
     "trip_title",
     "trip_destination",
     "invitee_email"
    ],
    "title": "InvitationResponse"
   },
...
   "MemberResponse": {
    "properties": {
     "user_id": {
      "type": "integer",
      "title": "User Id"
     },
     "name": {
      "type": "string",
      "title": "Name"
     },
     "email": {
      "type": "string",
      "format": "email",
      "title": "Email"
     },
     "role": {
      "type": "string",
      "title": "Role"
     },
     "avatar_url": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Avatar Url"
     }
    },
    "type": "object",
    "required": [
     "user_id",
     "name",
     "email",
     "role"
    ],
    "title": "MemberResponse"
   },
   "MemberRoleUpdate": {
    "properties": {
     "role": {
      "type": "string",
      "enum": [
       "owner",
       "member",
       "viewer"
      ],
      "title": "Role"
     }
    },
    "type": "object",
    "required": [
     "role"
    ],
    "title": "MemberRoleUpdate"
   },
...
   "TripResponse": {
    "properties": {
     "id": {
      "type": "integer",
      "title": "Id"
     },
     "title": {
      "type": "string",
      "title": "Title"
     },
     "description": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Description"
     },
     "destination": {
      "type": "string",
      "title": "Destination"
     },
     "destinations": {
      "items": {
       "$ref": "#/components/schemas/Destination"
      },
      "type": "array",
      "title": "Destinations",
      "default": []
     },
     "start_date": {
      "anyOf": [
       {
        "type": "string",
        "format": "date"
       },
       {
        "type": "null"
       }
      ],
      "title": "Start Date"
     },
     "end_date": {
      "anyOf": [
       {
        "type": "string",
        "format": "date"
       },
       {
        "type": "null"
       }
      ],
      "title": "End Date"
     },
     "budget": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Budget"
     },
     "currency": {
      "type": "string",
      "title": "Currency"
     }
    },
    "type": "object",
    "required": [
     "id",
     "title",
     "destination",
     "currency"
    ],
    "title": "TripResponse"
   },
...
   "TripUpdate": {
    "properties": {
     "title": {
      "anyOf": [
       {
        "type": "string",
        "maxLength": 255
       },
       {
        "type": "null"
       }
      ],
      "title": "Title"
     },
     "description": {
      "anyOf": [
       {
        "type": "string",
        "maxLength": 5000
       },
       {
        "type": "null"
       }
      ],
      "title": "Description"
     },
     "destination": {
      "anyOf": [
       {
        "type": "string",
        "maxLength": 255
       },
       {
        "type": "null"
       }
      ],
      "title": "Destination"
     },
     "destinations": {
      "anyOf": [
       {
        "items": {
         "$ref": "#/components/schemas/Destination"
        },
        "type": "array",
        "maxItems": 10,
        "minItems": 1
       },
       {
        "type": "null"
       }
      ],
      "title": "Destinations"
     },
     "start_date": {
      "anyOf": [
       {
        "type": "string",
        "format": "date"
       },
       {
        "type": "null"
       }
      ],
      "title": "Start Date"
     },
     "end_date": {
      "anyOf": [
       {
        "type": "string",
        "format": "date"
       },
       {
        "type": "null"
       }
      ],
      "title": "End Date"
     },
     "budget": {
      "anyOf": [
       {
        "type": "number",
        "maximum": 9999999999.0,
        "minimum": 0.0
       },
       {
        "type": "null"
       }
      ],
      "title": "Budget"
     },
     "currency": {
      "anyOf": [
       {
        "type": "string",
        "maxLength": 3,
        "minLength": 3
       },
       {
        "type": "null"
       }
      ],
      "title": "Currency"
     }
    },
    "type": "object",
    "title": "TripUpdate"
   },
...
   "ValidationError": {
    "properties": {
     "loc": {
      "items": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "integer"
        }
       ]
      },
      "type": "array",
      "title": "Location"
     },
     "msg": {
      "type": "string",
      "title": "Message"
     },
     "type": {
      "type": "string",
      "title": "Error Type"
     },
     "input": {
      "title": "Input"
     },
     "ctx": {
      "type": "object",
      "title": "Context"
     }
    },
    "type": "object",
    "required": [
     "loc",
     "msg",
     "type"
    ],
    "title": "ValidationError"
   },
...
  },
  "securitySchemes": {
   "HTTPBearer": {
    "type": "http",
    "scheme": "bearer"
   }
  }
 }
}